change: styles and overlays
