+++ lib/components/WideImage/WideImage.stories.tsx
@@ -3,7 +3,6 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { WideImage } from './WideImage'
 import { Button } from '../Button/Button'
 import { ButtonGroup } from '../ButtonGroup/ButtonGroup'
-import { ChevronDownIcon } from '@heroicons/react/24/solid'
 
 const meta: Meta<typeof WideImage> = {
   title: 'Components/Wide Image',
@@ -22,14 +21,6 @@ export const Default: Story = {}
 
 Default.args = {
   title: 'Wide image default',
-}
-
-const handleScroll = () => {
-  window.scroll({
-    top: document.body.offsetHeight,
-    left: 0,
-    behavior: 'smooth',
-  })
 }
 
 export const WithButtons: Story = {
@@ -55,12 +46,6 @@ export const BgDark: Story = {
         <Button title="Apply Now" />
         <Button title="Request Information" color="grey" />
       </ButtonGroup>
-      <ChevronDownIcon
-        onClick={handleScroll}
-        className="w-8 h-8 mt-5 transition-transform rotate-0 accordion__icon flex items-center justify-center cursor-pointer"
-        stroke={'white'}
-        strokeWidth={2}
-      />
     </WideImage>
   ),
 }
@@ -90,16 +75,17 @@ BgImage.args = {
 export const ScrollTo: Story = {
   render: (args) => (
     <WideImage {...args}>
-      <p>
-        Nobis voluptatem dolorum et eum doloremque cupiditate velit. Praesentium architecto a distinctio aut
-        reprehenderit ducimus.
-      </p>
+      <p>Author Name Goes Here</p>
+      <p>Add a Photo Credit</p>
     </WideImage>
   ),
 }
 ScrollTo.args = {
   ...Default.args,
+  maxHeight: 'lg',
+  scrollTo: true,
   title: 'Wide image with bg image',
+  headerType: 'h1',
   isType: 'image',
   image: 'https://picsum.photos/1600/700',
 }
